消除全部编译警告:清理未使用变量/字段/事件
- CANInsight.Core: 删除未用字段与局部变量(refersh/action/timesamp等),
  修复Recorder.cs CS0675符号扩展按位或, catch块去除未用异常变量
- 主应用: 清理CanSend/ChartFrom/ChartControl/LogFileToCSV/Main/
  PptReportBuilder中的未用字段与变量, catch块去未用ex
- Lin_API: 激活从未触发的LinkLost事件(OnLinkLost入口通知UI),
  使其与已有UI订阅者配合生效
- ProbeSmoke: 删除gone13死循环检查

验证: 主工程/CANInsight.Service/ProbeSmoke 全部0警告0错误

+++ CANInsight.Core/CAN_Data/Recorder.cs
@@ -237,7 +237,7 @@ namespace PCAN_Client.CAN_Data
                     var binWriter = new List<byte>();
                     binWriter.Add((byte)(copyLen + 6));
                     // ID高字节的高5位携带通道号（ID≤0x7FF只用低3位）：旧文件无通道信息时读出0→归一化为通道1，向后兼容
-                    binWriter.Add((byte)((msg.Channel << 3) | ((msg.CanId >> 8) & 0x07)));
+                    binWriter.Add((byte)(((uint)msg.Channel << 3) | ((msg.CanId >> 8) & 0x07)));
                     binWriter.Add((byte)(msg.CanId & 0xFF));
                     binWriter.Add((byte)(delta >> 16));
                     binWriter.Add((byte)(delta >> 8));
